Add function for board to identify valid moves
The function, find_valid_moves identifies the available positions to place a disk of the given color. 

For each disk of the given color, we check if their adjacent disks are of the other color to determine if it is possible to make a move. Then apply_move decides if the move is valid and executes it.