Count tiles on game-board method
Method that traverses through the grid to find the exact number of 
white, black and free spaces on the game board.

The method traverses through the grid and inspects each tile 
separately. Each tile will increment one of the three categories, and 
once the traversal is complete, a list of the category totals will 
return.