Add the get_move method to player
The method get_move calls a few functions to apply the move made by the player. 

We identify the available positions of the given color using find_valid_moves and assign it to validMoves. We call the function get_mouse_input to get the exact position clicked by the mouse on the board and if the move is valid, we terminate the current loop and call apply_move to finally execute the move.