Add the title screen graphics of Flipsies
This commit adds the initial screen that is displayed when you launch the 
game. The screen contains a title, a play button, an instructions button, 
and a logo design that resembles two overlapping game pieces.

A title screen is an expected convention of online games. The minimalist 
user interface ensures that players will intuitively understand how to 
navigate to the game.