Add method to identify valid moves
This method identifies the available squares on which 
a player may place a disk of the given color. It is in 
the board class.

For each disk of the given color, we check if their adjacent 
disks are of the other color to determine if the move 
adheres to the rules of Flipsies.